requirements.txt
openai
python-dotenv
supabase
flask
PyPDF2
requests
pdfminer.six